pyproject.toml
[tool.poetry]
name = "pytestifypro"
version = "0.1.0"
description = "Python testing framework that enhances pytest by offering utility functions and streamlined configurations. It simplifies writing, executing, and managing tests, making it easier to achieve robust and reliable testing outcomes."
authors = ["Javed Ansari <javed.ansari@strongholdindia.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.0"
requests = "^2.32.3"
pytest-mock = "^3.14.0"
jsonschema = "^4.23.0"
pyyaml = "^6.0.2"
allure-pytest = "^2.13.5"
selenium = "^4.0.0"
psycopg2 = "^2.9.0"
pytest-bdd = "^5.0.0"


#jaraco.functools = "^4.0.2"  #Poetry is having issues with this hence commenting for now until finds a fix
allure-python-commons = "^2.13.5"


[tool.poetry.dev-dependencies]
pytest-cov = "^5.0.0"
pylint = "^3.2.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


